handled pagebation , search , delete category edite and view the category image

+++ src/app/[locale]/dashboard/categories/_actions/actions.ts
@@ -1,13 +1,24 @@
 "use server";
 
-const token = `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoiNjhlZTA3YWY3ZmVlNjhhNGMyZWJhZmJhIiwicm9sZSI6InVzZXIiLCJpYXQiOjE3NjI0MjQzMDl9.0tYKHrz6liDX-0U_XLsrroB7ISnmChOip5evszIyiag`;
+import { Category } from "@app/[locale]/(home)/products/_types/categories";
+import { getDecodeToken } from "@lib/utils/get-decode-token";
+import { revalidateTag } from "next/cache";
+
+export interface UpdateCategoryResponse {
+    message: string;
+    category: Category;
+}
 
 export async function createCategoryAction(formData: FormData) {
+    const decodedToken = await getDecodeToken();
+    if (!decodedToken) {
+        throw new Error("No valid authentication token found");
+    }
     const res = await fetch("https://flower.elevateegy.com/api/v1/categories", {
         method: "POST",
         body: formData, // do not set Content-Type manually
         headers: {
-            Authorization: `Bearer ${token}`, // fine for testing
+            Authorization: `Bearer ${decodedToken.accessToken}`,
         },
         //revalidate tags
         next: { tags: ["categories"] },
@@ -16,23 +27,52 @@ export async function createCategoryAction(formData: FormData) {
     if (!res.ok) {
         console.error("Failed to create category");
     }
+    //revalidate pathes
+    revalidateTag("categories");
 
     return res.json();
 }
 
+export async function updateCategoryAction(categoryId: string, formData: FormData) {
+    const decodedToken = await getDecodeToken();
+    if (!decodedToken) {
+        throw new Error("No valid authentication token found");
+    }
+    const res = await fetch(`https://flower.elevateegy.com/api/v1/categories/${categoryId}`, {
+        method: "PUT",
+        body: formData, // do not set Content-Type manually
+        headers: {
+            Authorization: `Bearer ${decodedToken.accessToken}`, // fine for testing
+        },
+    });
+
+    if (!res.ok) {
+        console.error("Failed to update category");
+    }
+
+    const payload: UpdateCategoryResponse = await res.json();
+    //revalidate pathes
+    revalidateTag("categories");
+
+    return payload;
+}
+
 export async function deleteCategoryAction(categoryId: string) {
+    const decodedToken = await getDecodeToken();
+    if (!decodedToken) {
+        throw new Error("No valid authentication token found");
+    }
     const res = await fetch(`https://flower.elevateegy.com/api/v1/categories/${categoryId}`, {
         method: "DELETE",
         headers: {
-            Authorization: `Bearer ${token}`, // fine for testing
+            Authorization: `Bearer ${decodedToken.accessToken}`, // fine for testing
         },
-        //revalidate pathes
-        next: { tags: ["categories"] },
     });
 
     if (!res.ok) {
         console.error("Failed to delete category");
     }
-
+    //revalidate pathes
+    revalidateTag("categories");
     return res.json();
 }
